tidy includes

+++ include/surface_normal.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <vector>
-
 #include <opencv2/core.hpp>
 
 using Plane = cv::Vec4f;
@@ -18,6 +16,6 @@ cv::Mat3f normals_from_depth(const cv::Mat1f &depth, CameraIntrinsics intrinsics
 cv::Mat3b normals_to_rgb(const cv::Mat3f &normals);
 
 cv::Mat1f get_surrounding_points(const cv::Mat1f &depth, int i, int j, CameraIntrinsics intrinsics,
-                                 size_t window_size, float max_rel_depth_diff);
+                                 int window_size, float max_rel_depth_diff);
 
 cv::Vec3f fit_plane(const cv::Mat &points);
